mange permissions in a custom hook
we need to show the appStatusModal whenever the overallStatus is false

now that the controls are split from the modal, we need to have central access to the overallStatus

usePermissionStatus controls all the permissions from a central location. overallStatus can be easily extracted to show the modal in AppStatusModal

in PermissionsControls, we now extract the checkList, overallStatus, etc from usePermissionStatus in order to control the UI
