+++ www/js/control/AppStatusModal.tsx
@@ -1,12 +1,28 @@
-import React, { useState, useEffect, useMemo } from "react";
-import { Modal,  useWindowDimensions, ScrollView, View } from "react-native";
-import { Dialog, Button, Text, useTheme } from 'react-native-paper';
+import React, { useEffect } from "react";
+import { Modal,  useWindowDimensions } from "react-native";
+import { Dialog  } from 'react-native-paper';
 import PermissionsControls from "../appstatus/PermissionsControls";
+import usePermissionStatus from "../usePermissionStatus";
 //TODO -- import settings styles for dialog
 
 const AppStatusModal = ({permitVis, setPermitVis, dialogStyle}) => {
-    const {colors} = useTheme();
     const { height: windowHeight } = useWindowDimensions();
+    const { overallStatus } = usePermissionStatus();
+
+    //anytime the status changes, may need to show modal
+    useEffect(() => {
+        let currentlyOpen = window?.appStatusModalOpened;
+        if(!currentlyOpen && overallStatus == false) { //trying to block early cases from throwing modal
+            window.appStatusModalOpened = true;
+            setPermitVis(true);
+        }
+    }, [overallStatus]);
+
+    useEffect (() => {
+        if(!permitVis) {
+            window.appStatusModalOpened = false;
+        }
+    }, [permitVis]);
 
     return (
         <Modal visible={permitVis} onDismiss={() => setPermitVis(false)} transparent={true}>
@@ -14,7 +30,7 @@ const AppStatusModal = ({permitVis, setPermitVis, dialogStyle}) => {
                         onDismiss={() => setPermitVis(false)} 
                         style={dialogStyle}>
                     <Dialog.Content  style={{maxHeight: windowHeight/1.5, paddingBottom: 0}}>
-                            <PermissionsControls permitVis={permitVis} setPermitVis={setPermitVis}></PermissionsControls>
+                            <PermissionsControls onAccept={() => setPermitVis(false)}></PermissionsControls>
                     </Dialog.Content>
 
                 </Dialog>
